ensure default profile image is not delete

+++ service/userService.ts
@@ -10,6 +10,7 @@ import { ImageKeyStrategy } from "../helper/s3KeyStrategy.js"
 import { Role } from "../utils/enums.js";
 import { getValidRoles } from "../utils/roleUtils.js";
 import { createProfileStrategy } from "../helper/createProfileStrategy.js";
+import { DEFAULT_PROFILE_IMAGE_KEY } from "../utils/constants.js";
 
 
 
@@ -137,6 +138,13 @@ export class UserService {
         }
     }
 
+    async check_default_profile_image(profile_image_key: string){
+        if (profile_image_key === DEFAULT_PROFILE_IMAGE_KEY) {
+            return true;
+        }
+        return false;
+    }
+
     async update_profile_image(file: Express.Multer.File, user: IUserRequest){
         const existingUser = await this.userRepository.get_user_by_id(user.id);
         const oldProfileImage = existingUser.profile_image;
@@ -146,7 +154,9 @@ export class UserService {
         const newKey = await this.upload_profile_image(file, user);
         try{
             //delete old image from s3
-            await this.s3Service.deleteFile(oldProfileImage);
+            if (!(await this.check_default_profile_image(oldProfileImage))) {
+                await this.s3Service.deleteFile(oldProfileImage);
+            }
         }catch(error: unknown){
             console.error((error as Error).message);
             throw new Error("Failed to delete old profile image");
@@ -170,6 +180,9 @@ export class UserService {
         const user = await this.userRepository.get_user_by_id(user_id);
         if(!user.profile_image){
             throw new Error("No profile image found");
+        }
+        if(await this.check_default_profile_image(user.profile_image)){
+            throw new Error("Cannot delete default profile image");
         }
         try{
             await this.s3Service.deleteFile(user.profile_image);
